ai-engine/requirements.txt
spacy
pandas
numpy
scikit-learn
nltk
PyPDF2
python-docx
fastapi
uvicorn
pytest
